requirements.txt
python-dateutil == 2.8.2
types-python-dateutil == 1.8.15
backoff == 2.0.1
BeautifulSoup4 == 4.11.1
types-beautifulsoup4 == 4.11.1
requests == 2.27.1
types-requests == 2.27.25
pyinstaller == 5.0.1
Pillow == 9.1.0
types-Pillow == 9.0.15
